Moved datasets and file handlers to loader package.
