+++ pymepps/loader/filehandler/wmtexthandler.py
